+++ samples/inbox/src/commonMain/kotlin/com/slack/circuit/sample/inbox/home/InboxScreen.kt
@@ -8,11 +8,12 @@ import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.screen.Screen
 import com.slack.circuit.sample.inbox.detail.EmailDetailScreen
 import com.slack.circuit.sample.inbox.list.InboxListScreen
-import kotlinx.serialization.Serializable
+import com.slack.circuit.serialization.CircuitSerializable
+import dev.zacsweers.metro.AppScope
 
 /** Root screen for the adaptive inbox surface. */
 @Parcelize
-@Serializable
+@CircuitSerializable(AppScope::class)
 data object InboxScreen : Screen {
   data class State(
     val listState: InboxListScreen.State,