+++ samples/star/src/androidMain/kotlin/com/slack/circuit/star/MainActivity.kt
@@ -68,10 +68,16 @@ class MainActivity(private val circuit: Circuit) : AppCompatActivity() {
     setContent {
       val uriHandler = remember { CustomTabsUriHandler() }
       CompositionLocalProvider(LocalUriHandler provides uriHandler) {
-        val backStack = rememberSaveableBackStack(initialBackstack, circuitSaver = starCircuitSaver)
+        val circuitSaver = requireNotNull(circuit.circuitSaver)
+        val backStack = rememberSaveableBackStack(initialBackstack, circuitSaver = circuitSaver)
         val circuitNavigator = rememberCircuitNavigator(backStack)
         val navigator = rememberAndroidScreenAwareNavigator(circuitNavigator, this::goTo)
-        val state = rememberStarAppState(backStack = backStack, navigator = navigator)
+        val state =
+          rememberStarAppState(
+            circuitSaver = circuitSaver,
+            backStack = backStack,
+            navigator = navigator,
+          )
         StarCircuitApp(circuit, state = state)
       }
     }