+++ samples/star/src/commonMain/kotlin/com/slack/circuit/star/petlist/FiltersScreen.kt
@@ -19,15 +19,15 @@ import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuit.runtime.screen.PopResult
 import com.slack.circuit.runtime.screen.Screen
+import com.slack.circuit.serialization.CircuitSerializable
 import com.slack.circuit.star.common.BackPressNavIcon
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedFactory
 import dev.zacsweers.metro.AssistedInject
-import kotlinx.serialization.Serializable
 
 @Parcelize
-@Serializable
+@CircuitSerializable(AppScope::class)
 data class FiltersScreen(val initialFilters: Filters) : Screen {
   data class State(val initialFilters: Filters, val eventSink: (Event) -> Unit) : CircuitUiState
 
@@ -35,7 +35,9 @@ data class FiltersScreen(val initialFilters: Filters) : Screen {
     data class Save(val filters: Filters) : Event
   }
 
-  @Parcelize @Serializable data class Result(val filters: Filters) : PopResult
+  @Parcelize
+  @CircuitSerializable(AppScope::class)
+  data class Result(val filters: Filters) : PopResult
 }
 
 @AssistedInject